Adopt new privileged launch method
The framework determines whether to use a privileged ShipIt based on whether it can write to the target location, I tested that this can replace root:wheel owned test applications
The tests statically use a nonprivileged helper because Authorization prompts for username/password.

The username/password can be provided to the authorization API but we'd need to know them in advance, perhaps through environment variables. Other than that I’m not sure how to test this.