Break enumeration down into a signal of signals
Change `-updateFileSecurity:` to return a signal, this contains the get, modify and set of the file security and captures all the errors.

I still have the modification of the `CFFileSecurity` in a separate method because it is non error byref to avoid duplicating the error handling at each `return NO` call site.