Remove resumable downloads after one completes
The assumption is that all other pending downloads are now redundant because we have an update and have moved it out of the resumable download cache.

This will be skipped if there’s an error, and we ignore errors removing the other downloads, the caller only wants to know about errors downloading the update it wanted.