Replace `SQRLShipItState` with `SQRLShipItRequest`
This avoids the state parameters being passed to the root installer, and forces the root installer to store its state in a non-user accessible location.

This reimplements the the installer to instead of tracking the state of the update request, only track the state of the bundle that is being replaced.

Each time we launch we can prepare a fresh update.

This handles the problem where on launch we don’t know if we crashed/failed on a prior launch and are resuming an update, or have been given a new update request to handle. This means we don’t have to duplicate the incoming request into secure storage, because the incoming request doesn’t contain any of the installer state.